+++ odoo/local-src/connector_esb/controllers/sale.py
@@ -137,8 +137,8 @@ class SaleController(http.Controller):
         self._validate_request(request.jsonrequest)
         values = request.jsonrequest['params']['data']
         self._validate_create_sale_order(values)
-        delayable = env['sale.order'].with_delay(priority=25)
-        delayable.ws_create_new(values)
+        delayable = env['sale.order'].with_delay(priority=2)
+        delayable.ws_create_new(values, datetime.now())
 
     @http.route(
         '/connector_esb/sales_order/status',