+++ HISTORY.rst
@@ -21,9 +21,13 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* ALCYN-2313: change priority of job "Create a sale order with data coming from
+  webservices" to 2 instead of 25
+
 **Bugfixes**
 
 * ALCYN-2317: fix issues with purchase discounts
+* ALCYN-2309: Fix cancellation of sale orders after some delay
 
 **Build**